requirements.txt
fastapi
uvicorn
pytest
httpx
pyjwt